Phase 1/Simulator/requirements.txt
packaging==25.0
pymodbus==3.11.1
PyYAML==6.0.2
Flask
pyModbusTCP
